refactor: move JWT_KEY to constants
