--- src/config/constant.ts
@@ -1,1 +1,0 @@
-export const JWT_KEY = 'jwt';
